Fixed IHasLongTotalCount.TotalCount as long.

+++ src/Abp/Application/Services/Dto/IHasLongTotalCount.cs
@@ -8,6 +8,6 @@ namespace Abp.Application.Services.Dto
         /// <summary>
         /// Total count of Items.
         /// </summary>
-        int TotalCount { get; set; }
+        long TotalCount { get; set; }
     }
 }
